Achieve a highly maintainable and context-driven architecture for your habits feature
