+++ src/features/habits/components/AddHabitDialog.tsx
@@ -1,5 +1,5 @@
 import HabitDialog from "@/features/habits/components/HabitDialog";
-import HabitDialogHeader from "@/features/habits/components/dialog/HabitDialogHeader";
+import HabitDialogHeader from "@/features/habits/components/HabitDialogHeader";
 import HabitForm from "@/features/habits/components/HabitForm";
 import { useHabitsState } from "@/features/habits/context/HabitsStateContext";
