+++ src/features/habits/components/EditHabitDialog.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import HabitDialog from "@/features/habits/components/HabitDialog";
-import HabitDialogHeader from "@/features/habits/components/dialog/HabitDialogHeader";
+import HabitDialogHeader from "@/features/habits/components/HabitDialogHeader";
 import HabitForm from "@/features/habits/components/HabitForm";
 import HabitStats from "@/features/habits/components/HabitStats";
 import { ThemeContext } from "@/context/ThemeContext";
